Refactor blog creation and editing components
Consolidated blog creation and editing logic into a reusable `FormBlogComponent` to reduce redundancy and improve maintainability. Updated related routes, types, Redux slices, and UI components to support the new structure and added API patch support for editing blogs.
